+++ src/main/java/controller/UserController.java
@@ -1,18 +1,19 @@
 package controller;
 
-import annotation.MyRequestMapping;
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
 import com.github.jknack.handlebars.io.TemplateLoader;
 import db.DataBase;
+import framework.annotation.MyRequestMapping;
+import framework.controller.Controller;
+import framework.request.HttpCookie;
+import framework.request.Request;
+import framework.response.ContentType;
+import framework.response.Response;
+import framework.utils.FileIoUtils;
 import model.User;
-import request.HttpCookie;
-import request.Request;
-import response.ContentType;
-import response.Response;
 import service.Session;
 import service.SessionHandler;
-import utils.FileIoUtils;
 
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
