+++ src/main/java/framework/WebServer.java
@@ -1,4 +1,4 @@
-package webserver;
+package framework;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
