+++ src/main/java/framework/annotation/MyRequestMapping.java
@@ -1,4 +1,4 @@
-package annotation;
+package framework.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
