+++ src/main/java/framework/request/HttpCookie.java
@@ -1,4 +1,4 @@
-package request;
+package framework.request;
 
 import java.util.HashMap;
 import java.util.Map;
