+++ src/main/java/framework/response/ContentType.java
@@ -1,4 +1,4 @@
-package response;
+package framework.response;
 
 public enum ContentType {
     HTML("text/html;charset=utf-8"),
