+++ src/main/java/framework/utils/IOUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package framework.utils;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
