+++ src/test/java/utils/IOUtilsTest.java
@@ -1,5 +1,6 @@
 package utils;
 
+import framework.utils.IOUtils;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
