+++ src/test/java/webserver/RequestHandlerTest.java
@@ -1,9 +1,10 @@
 package webserver;
 
 import db.DataBase;
+import framework.RequestHandler;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
-import utils.FileIoUtils;
+import framework.utils.FileIoUtils;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
